% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{display_plot_code_module_server}
\alias{display_plot_code_module_server}
\title{Display the plotting code in server}
\usage{
display_plot_code_module_server(id, plot_string)
}
\description{
Display the plotting code in server
}
