% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{display_plot_code_module_ui}
\alias{display_plot_code_module_ui}
\title{Display the plotting code in UI}
\usage{
display_plot_code_module_ui(id, width = "600px", height = "400px")
}
\description{
Display the plotting code in UI
}
