% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{rowCallback}
\alias{rowCallback}
\title{Thanks, stackoverflow!
https://stackoverflow.com/questions/58526047/customizing-how-datatables-displays-missing-values-in-shiny}
\format{
An object of class \code{character} of length 8.
}
\usage{
rowCallback
}
\description{
Thanks, stackoverflow!
https://stackoverflow.com/questions/58526047/customizing-how-datatables-displays-missing-values-in-shiny
}
\keyword{datasets}
